requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
